\name{print.mrqap.dsp}
\alias{print.mrqap.dsp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print function for mrqap.dsp
}
\description{
Print function for MRQAP DSP function
}
\usage{
\method{print}{mrqap.dsp}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an mrqap.dsp object (from function in asnipe package)
}
  \item{\dots}{ Further arguments passed to or from print methods }
}
\details{
Print formatted results.
}
\value{
Prints formatted results
}
\references{
see mrqap.dsp function
}
\author{
Damien R. Farine
}

