\name{asnipe-package}
\alias{asnipe-package}
\alias{asnipe}
\docType{package}
\title{
Animal Social Network inference and Permutation: asnipe
}
\description{
Provides functions for infering associations, building social networks, performing permutations, and regression testing
}
\details{
\tabular{ll}{
Package: \tab asnipe\cr
Type: \tab Package\cr
Version: \tab 0.90\cr
Date: \tab 2016-04-03\cr
License: \tab GPL-2 \cr
}
}
\author{
Written by Damien R. Farine

Maintainer: Damien R. Farine <dfarine@orn.mpg.de>
}
